+++ moor/lib/src/runtime/query_builder/expressions/null_check.dart
@@ -4,11 +4,24 @@ part of '../query_builder.dart';
 // look together with NNBD in the future
 
 /// Expression that is true if the inner expression resolves to a null value.
+@Deprecated('Use isNull through the SqlIsNull extension')
 Expression<bool> isNull(Expression inner) => _NullCheck(inner, true);
 
 /// Expression that is true if the inner expression resolves to a non-null
 /// value.
+@Deprecated('Use isNotNull through the SqlIsNull extension')
 Expression<bool> isNotNull(Expression inner) => _NullCheck(inner, false);
+
+/// Extension defines the `isNull` and `isNotNull` members to check whether the
+/// expression evaluates to null or not.
+extension SqlIsNull on Expression {
+  /// Expression that is true if the inner expression resolves to a null value.
+  Expression<bool> isNull() => _NullCheck(this, true);
+
+  /// Expression that is true if the inner expression resolves to a non-null
+  /// value.
+  Expression<bool> isNotNull() => _NullCheck(this, false);
+}
 
 /// Evaluates to the first expression in [expressions] that's not null, or
 /// null if all [expressions] evaluate to null.